Re-introduce lambda functions
Turns out they *were* necessary.
If they're not used, the same "instance" of the dialog box is used. For example,
when the SysInfo box is closed and you click the button again, it shows an already
filled-out question sheet.
This can probably be done in some better way, but this works